Improve potential divider voltage calculating
The function gets a multiplying parameter (mltpty) that was supposed to increase the integer calculations of the function.

We multiply the intermediate output and than divide it back.
But for the case of calc_type_vin we don’t divide it so this is a trap (because now the result is a multiplied result).

So the quick fix is to divide the multiplier back