Fix bug introduced in previous release that caused deviceInstance to disappear.
The previous version made an attempt to not start services unless the settings
say that we must. This is already done on the first sensor read: All sensors
that should be connected but aren't, will be connected, and vice versa. The clincher
is that this is also how we handle sensors that are enabled AFTER startup via
the settings. It is therefore acceptable to simply not enable anything at startup
and let it happen on the first clock tick.

This avoids the original problem, where the service appeared and disappeared
really quickly, leaving behind a dead entry in the GUI. But this time we don't
lose our deviceInstance.