allow custom tank resistance settings
Make the tank empty and tank full resistance configurable. Do make sure
these are kept in sync with the predefined ones, they are restored when
the standard is no set to custom. For normal cases the gui should enforce
this. Do allow the tank empty resistance to be higer then the tank low,
and merge the calculation for the USA vs EU standard.