set empty/full items only when value is changed
In updateTankLevels(), set the velib items only if there has been a
change. This avoids a situation where the value is forwarded to
localsettings which then sends it back, potentially overwriting a change
that has occurred in the meantime.

This fixes a weird rapid toggling between two values seen after changing
several settings quickly via MQTT.

https://github.com/victronenergy/venus-private/issues/352